privatedns: refactoring to use the new structure
